user registration added

+++ blogsite/blogs/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render,redirect
 from django.db.models import Q
+from django.http import HttpResponse
 from .models import blog,topic,User
 from .forms import blogform,topicform
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 def index(request):
     return render(request,'index.html')
@@ -27,19 +30,26 @@ def blogpage(request,slug):
     blogpost = blog.objects.get(slug=slug)
     return render(request, 'blogpage.html', {'blogpost':blogpost})
 
+@login_required(login_url='/login')
 def createblog(request):
     if request.method == 'POST':
         blog = blogform(request.POST)
         if blog.is_valid():
             obj = blog.save(commit=False)
             obj.slug = blog.data['title'].replace(' ','-')
+            obj.author = request.user
             obj.save()
             return redirect('index')
     return render(request,'addblog.html',{'form':blogform})
 
+@login_required(login_url='/login')
 def updateblog(request, id):
     blogpost = blog.objects.get(id=id)
     form = blogform(instance=blogpost)
+
+    if(blogpost.author!=request.user.username):
+        return HttpResponse("You're not allowed to be here!")
+
     if(request.method=='POST'):
         newform = blogform(request.POST, instance=blogpost)
         if newform.is_valid():
@@ -47,16 +57,20 @@ def updateblog(request, id):
             return redirect('blogs')
     return render(request,'updateblog.html',{'form':form})
 
+@login_required(login_url='/login')
 def deleteblog(request,id):
     blogpost = blog.objects.get(id=id)
+    if(blogpost.author!=request.user.username):
+        return HttpResponse("You're not allowed to be here!")
     if request.method=='POST':
         blogpost.delete()
         return redirect('blogs')
     return render(request,'delete.html',{'obj':blogpost})
 
 def loginpage(request):
+    page = 'login'
     if request.method=='POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -68,12 +82,25 @@ def loginpage(request):
             return redirect('index')
         else:
             messages.error(request,"Username/Password is wrong!")
-    return render(request,'login_register.html')
+    return render(request,'login_register.html',{'page':page})
 
+def register(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user=user)
+            return redirect('index')
+    return render(request,'login_register.html',{'form':form})
 def logoutuser(request):
     logout(request)
     return redirect('index')
 
+@login_required(login_url='/login')
 def addtopic(request):
     if request.method=='POST':
         topic = topicform(request.POST)
